move Tech info out
