+++ expo/gbe/models/remains.py
@@ -40,6 +40,7 @@ from gbe.models import (
     Persona,
     Profile,
     Room,
+    TechInfo,
     VolunteerWindow,
 )
 
@@ -85,279 +86,6 @@ class Biddable(models.Model):
     def is_current(self):
         return self.conference.status in ("upcoming", "current")
 
-
-###################
-# Technical info #
-###################
-
-class AudioInfo(models.Model):
-    '''
-    Information about the audio required for a particular Act
-    '''
-    track_title = models.CharField(max_length=128, blank=True)
-    track_artist = models.CharField(max_length=123, blank=True)
-    track = models.FileField(upload_to='uploads/audio', blank=True)
-    track_duration = DurationField(blank=True)
-    need_mic = models.BooleanField(default=False, blank=True)
-    own_mic = models.BooleanField(default=False, blank=True)
-    notes = models.TextField(blank=True)
-    confirm_no_music = models.BooleanField(default=False)
-
-    @property
-    def dump_data(self):
-        return [self.track_title,
-                self.track_artist,
-                self.track,
-                self.track_duration,
-                self.need_mic,
-                self.own_mic,
-                self.notes,
-                self.confirm_no_music
-                ]
-
-    def clone(self):
-        ai = AudioInfo(track_title=self.track_title,
-                       track_artist=self.track_artist,
-                       track=self.track,
-                       track_duration=self.track_duration,
-                       need_mic=self.need_mic,
-                       own_mic=self.own_mic,
-                       notes=self.notes,
-                       confirm_no_music=self.confirm_no_music)
-        ai.save()
-        return ai
-
-    @property
-    def is_complete(self):
-        return bool(self.confirm_no_music or
-                    (self.track_title and
-                     self.track_artist and
-                     self.track_duration
-                     ))
-
-    @property
-    def incomplete_warnings(self):
-        if self.is_complete:
-            return {}
-        else:
-            return {'audio': audioinfo_incomplete_warning}
-
-    def __unicode__(self):
-        try:
-            return "AudioInfo: " + self.techinfo.act.title
-        except:
-            return "AudioInfo: (deleted act)"
-
-    class Meta:
-        verbose_name_plural = 'audio info'
-        app_label = "gbe"
-
-
-class LightingInfo (models.Model):
-    '''
-    Information about the basic (not related to cues) lighting
-    needs of a particular Act
-    '''
-    notes = models.TextField(blank=True)
-    costume = models.TextField(blank=True)
-    specific_needs = models.TextField(blank=True)
-
-    def clone(self):
-        li = LightingInfo(notes=self.notes,
-                          costume=self.costume)
-        li.save()
-        return li
-
-    @property
-    def dump_data(self):
-        return [self.notes, self.costume]
-
-    @property
-    def is_complete(self):
-        return True
-
-    @property
-    def incomplete_warnings(self):
-        if self.is_complete:
-            return {}
-        else:
-            return {"lighting": lightinginfo_incomplete_warning}
-
-    def __unicode__(self):
-        try:
-            return "LightingInfo: " + self.techinfo.act.title
-        except:
-            return "LightingInfo: (deleted act)"
-
-    class Meta:
-        verbose_name_plural = 'lighting info'
-        app_label = "gbe"
-
-
-class StageInfo(models.Model):
-    '''
-    Information about the stage requirements for a particular Act
-    confirm field should be offered if the user tries to save with all
-    values false and no notes
-    '''
-    act_duration = DurationField(blank=True)
-    intro_text = models.TextField(blank=True)
-    confirm = models.BooleanField(default=False)
-    set_props = models.BooleanField(default=False)
-    cue_props = models.BooleanField(default=False)
-    clear_props = models.BooleanField(default=False)
-    notes = models.TextField(blank=True)
-
-    def clone(self):
-        si = StageInfo(act_duration=self.act_duration,
-                       intro_text=self.intro_text,
-                       confirm=self.confirm,
-                       set_props=self.set_props,
-                       cue_props=self.cue_props,
-                       clear_props=self.clear_props,
-                       notes=self.notes)
-        si.save()
-        return si
-
-    @property
-    def dump_data(self):
-        return [self.act_duration,
-                self.intro_text.encode('utf-8').strip(),
-                self.confirm,
-                self.set_props,
-                self.cue_props,
-                self.clear_props,
-                self.notes.encode('utf-8').strip(),
-                ]
-
-    @property
-    def is_complete(self):
-        return bool(self.set_props or
-                    self.clear_props or
-                    self.cue_props or self.confirm)
-
-    @property
-    def incomplete_warnings(self):
-        if self.is_complete:
-            return {}
-        else:
-            return {'stage': stageinfo_incomplete_warning}
-
-    def __unicode__(self):
-        try:
-            return "StageInfo: " + self.techinfo.act.title
-        except:
-            return "StageInfo: (deleted act)"
-
-    class Meta:
-        verbose_name_plural = 'stage info'
-        app_label = "gbe"
-
-
-class TechInfo(models.Model):
-    '''
-    Gathers up technical info about an act in a show.
-    BB - doing 2 additional cues for now for an easy add to existing DB
-      2015  - may want to consider a many to many field for 1+ cues
-    '''
-    audio = models.OneToOneField(AudioInfo, blank=True)
-    lighting = models.OneToOneField(LightingInfo, blank=True)
-    stage = models.OneToOneField(StageInfo, blank=True)
-
-    def clone(self):
-        ti = TechInfo()
-        ti.audio = self.audio.clone()
-        ti.lighting = self.lighting.clone()
-        ti.stage = self.stage.clone()
-        ti.save()
-        for ci in CueInfo.objects.filter(techinfo=self):
-            ci.clone(self)
-        return ti
-
-    @property
-    def is_complete(self):
-        try:
-            CueInfo.objects.get(techinfo=self, cue_sequence=0)
-            cueinfopresent = True
-        except CueInfo.DoesNotExist:
-            cueinfopresent = False
-        return bool(self.audio.is_complete and
-                    self.lighting.is_complete and
-                    self.stage.is_complete and
-                    cueinfopresent)
-
-    def get_incomplete_warnings(self):
-        warnings = {}
-        warnings.update(self.lighting.incomplete_warnings)
-        warnings.update(self.audio.incomplete_warnings)
-        warnings.update(self.stage.incomplete_warnings)
-        return warnings
-
-    def __unicode__(self):
-        try:
-            return "Techinfo: " + self.act.title
-        except:
-            return "Techinfo: (deleted act)"
-
-    class Meta:
-        verbose_name_plural = 'tech info'
-        app_label = "gbe"
-
-
-class CueInfo(models.Model):
-    '''
-    Information about the lighting needs of a particular Act as they
-    relate to one or more cues within the Act.  Each item is the change
-    that occurs after a cue
-    '''
-    cue_sequence = models.PositiveIntegerField(default=0)
-    cue_off_of = models.TextField()
-
-    follow_spot = models.CharField(max_length=25,
-                                   choices=follow_spot_options,
-                                   default=follow_spot_options[0])
-
-    center_spot = models.CharField(max_length=20,
-                                   choices=offon_options, default="OFF")
-
-    backlight = models.CharField(max_length=20,
-                                 choices=offon_options, default="OFF")
-
-    cyc_color = models.CharField(max_length=25,
-                                 choices=cyc_color_options,
-                                 default=cyc_color_options[0])
-
-    wash = models.CharField(max_length=25,
-                            choices=stage_lighting_options,
-                            default=stage_lighting_options[0])
-    sound_note = models.TextField(blank=True)
-    techinfo = models.ForeignKey(TechInfo)
-
-    def clone(self, techinfo):
-        CueInfo(cue_sequence=self.cue_sequence,
-                cue_off_of=self.cue_off_of,
-                follow_spot=self.follow_spot,
-                center_spot=self.center_spot,
-                backlight=self.backlight,
-                cyc_color=self.cyc_color,
-                wash=self.wash,
-                sound_note=self.sound_note,
-                techinfo=techinfo).save()
-
-    @property
-    def is_complete(self):
-        return bool(self.cue_off_of and self.cue_sequence and self.tech_info)
-
-    def __unicode__(self):
-        try:
-            return "%s - cue %s" % (self.techinfo.act.title,
-                                    str(self.cue_sequence))
-        except:
-            return "Cue: (deleted act) - %s" % str(self.cue_sequence)
-
-    class Meta:
-        verbose_name_plural = "cue info"
-        app_label = "gbe"
 
 
 #######
